Removes RawTextWriter
