+++ src/text/mod.rs
@@ -2,6 +2,6 @@ pub mod non_blocking;
 pub(crate) mod parent_container;
 pub(crate) mod parse_result;
 pub(in crate::text) mod parsers;
-pub(crate) mod raw_text_writer;
 pub(crate) mod text_formatter;
 pub(crate) mod text_value;
+pub(crate) mod whitespace_config;
